requirements.txt
aioredis==1.2.0
amqp==2.1.4
asgi-redis==1.4.0
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
billiard==3.5.0.2
bleach==2.0.0
celery==4.0.2
certifi==2017.4.17
cffi==1.12.2
channels==1.1.3
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==1.2.0
dj-database-url==0.5.0
Django==1.11
django-celery-beat==1.0.1
django-channels-presence==0.0.7
django-cors-headers==2.5.2
django-heroku==0.3.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
hiredis==1.0.0
html5lib==0.999999999
hyperlink==17.1.1
idna==2.5
incremental==17.5.0
kombu==4.0.2
Markdown==2.6.8
msgpack==0.5.6
msgpack-python==0.4.8
Pillow==5.4.1
postmarker==0.11.2
psycopg2==2.7.1
psycopg2-binary==2.7.4
pycparser==2.19
PyHamcrest==1.9.0
pytz==2017.2
redis==2.10.5
reportlab==3.5.18
requests==2.17.3
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
vine==1.1.3
webencodings==0.5.1
whitenoise==4.1.2
zope.interface==4.4.1
